RavenDB-22451 Fix rendering of custom sorters in server-wide list
